moved macros from puppet/parser/macros to puppet/macros
This breaks API but I decided to do this now, until it's an early
project. What users have to do is to just move their macros from
puppet/parser/macros to puppet/macros and autoloader will be able again
to find them.

The decision was made after the comment by Henrik Lindberg about new
(future) puppet functions. They (as well as our macros) are not part of
parser. See:
https://groups.google.com/forum/#!topic/puppet-dev/i89rxODz3G4